Fix PR gdb/20418 - Problems with synchronous commands and new-ui
When executing commands on a secondary UI running the MI interpreter,
some commands that should be synchronous are not.

Running the console interpreter on the extra UI instead does not
trigger the problem.  In the MI case, the MI ui's ui->prompt_state
does end up correctly set to PROMPT_BLOCKED.  The problem happens when
we emit MI async events (=library-loaded, etc.), and we go about
restoring the previous terminal state, we end up calling
target_terminal_ours, which reinstalls MI's input_fd in the event
loop...  Specifically, code like this:

   old_chain = make_cleanup_restore_target_terminal ();
   target_terminal_ours_for_output ();

   fprintf_unfiltered (mi->event_channel, "library-loaded");

...

   do_cleanups (old_chain);

The fix is to move the add_file_handler/delete_file_handler calls out
of target_terminal_foo, making the latter completly no-ops unless
called with the main UI as current UI.

Needs a test...