Add test for user context selection sync
This patch adds a test to verify that events are sent properly to all
UIs when the user selection context (inferior, thread, frame) changes.

The goal of the C test file is to provide two threads that are
interrupted with the same predictable backtrace (so that we can test
frame switching).  This is achieved by having them loop on a single
line, such that when the main thread hits a breakpoint, they are both
stopped that line.  It would not be practical to have the threads sleep,
since their backtraces would not be predictable (the functions that
implement sleep may vary between systems).

There is a 1 second sleep in the main thread to make sure the threads
have time to spawn and reach the loop.  If you can find a way that is
sleep-free and race-free to achieve the same result, it would be really
nice, as most of the time taken by the test is spent sleeping.

gdb/testsuite/ChangeLog:

YYYY-MM-DD  Antoine Tremblay  <antoine.tremblay@ericsson.com>
YYYY-MM-DD  Simon Marchi  <simon.marchi@ericsson.com>

	PR gdb/20487
	* gdb.mi/user-selected-context-sync.exp: New file.
	* gdb.mi/user-selected-context-sync.c: New file.