Make gdb_exception manage lifetime of its message string
Currently, the lifetime of the message string associated with
gdb_exception objects is managed by a stack on the side:

 /* A stack of exception messages.
    This is needed to handle nested calls to throw_it: we don't want to
    xfree space for a message before it's used.
    This can happen if we throw an exception during a cleanup:
    An outer TRY_CATCH may have an exception message it wants to print,
    but while doing cleanups further calls to throw_it are made.

    This is indexed by the size of the current_catcher list.
    It is a dynamically allocated array so that we don't care how deeply
    GDB nests its TRY_CATCHs.  */
 static char **exception_messages;

 /* The number of currently allocated entries in exception_messages.  */
 static int exception_messages_size;

This patch gets rid of that side stack by making the exception objects
responsible for managing the lifetime of their message strings.

To avoid introducing string copying whenever an exception object is
copied (currently we just copy the pointer), this patch makes use of
C++11 rval references, along with a simple emulation for C++03.
Instead of writting std::move, you write gdb::move, and instead of
writting "gdb_exception&&", you write gdb_exception_rval_ref.  The
latter is actually a struct in C++03 mode.  The C++03 mode is a bit
more permissive (I didn't want to complicate the code), but we know
that if the code compiles in C++11 mode (or later), then we're not
doing anything we shouldn't.  The next patch will make more use of
rvalue references throughout the codebase to eliminate unnecessary
gdb_exception::message xstrdup'ing introduced by this patch.